+++ src/ExceptionHandling/ExceptionHandlerInterface.php
@@ -14,8 +14,8 @@ declare(strict_types = 1);
 
 namespace SensioLabs\RichModelForms\ExceptionHandling;
 
+use Symfony\Component\Form\FormConfigInterface;
 use Symfony\Component\Form\FormError;
-use Symfony\Component\Form\FormInterface;
 
 /**
  * Converts exceptions into form errors.
@@ -32,5 +32,5 @@ use Symfony\Component\Form\FormInterface;
  */
 interface ExceptionHandlerInterface
 {
-    public function getError(FormInterface $form, $data, \Throwable $e): ?FormError;
+    public function getError(FormConfigInterface $formConfig, $data, \Throwable $e): ?FormError;
 }