+++ src/ExceptionHandling/GenericExceptionHandler.php
@@ -14,8 +14,8 @@ declare(strict_types = 1);
 
 namespace SensioLabs\RichModelForms\ExceptionHandling;
 
+use Symfony\Component\Form\FormConfigInterface;
 use Symfony\Component\Form\FormError;
-use Symfony\Component\Form\FormInterface;
 
 /**
  * A generic exception handler that can transform arbitrary exceptions into form errors.
@@ -38,7 +38,7 @@ final class GenericExceptionHandler implements ExceptionHandlerInterface
         $this->handledExceptionClass = $handledExceptionClass;
     }
 
-    public function getError(FormInterface $form, $data, \Throwable $e): ?FormError
+    public function getError(FormConfigInterface $formConfig, $data, \Throwable $e): ?FormError
     {
         if (!$e instanceof $this->handledExceptionClass) {
             return null;